Move global interface to tpbot package.
