+++ TelepresenceBot/core/src/main/java/com/novoda/tpbot/FeatureSelectionPersistence.java
@@ -1,4 +1,4 @@
-package com.novoda.tpbot.feature_selection;
+package com.novoda.tpbot;
 
 public interface FeatureSelectionPersistence {
