+++ TelepresenceBot/mobile/src/main/java/com/novoda/tpbot/bot/video/calling/VideoCallSharedPreferencesPersistence.java
@@ -3,7 +3,7 @@ package com.novoda.tpbot.bot.video.calling;
 import android.content.Context;
 import android.content.SharedPreferences;
 
-import com.novoda.tpbot.feature_selection.FeatureSelectionPersistence;
+import com.novoda.tpbot.FeatureSelectionPersistence;
 
 public final class VideoCallSharedPreferencesPersistence implements FeatureSelectionPersistence {
