+++ TelepresenceBot/mobile/src/main/java/com/novoda/tpbot/landing/LandingActivity.java
@@ -9,7 +9,7 @@ import android.view.View;
 
 import com.novoda.tpbot.R;
 import com.novoda.tpbot.bot.BotActivity;
-import com.novoda.tpbot.feature_selection.FeatureSelectionController;
+import com.novoda.tpbot.FeatureSelectionController;
 import com.novoda.tpbot.landing.menu.LandingMenuFeatureSelectionController;
 import com.novoda.tpbot.human.HumanActivity;
