+++ TelepresenceBot/mobile/src/main/java/com/novoda/tpbot/landing/menu/LandingMenuFeatureSelectionController.java
@@ -10,8 +10,8 @@ import android.view.MenuItem;
 import com.novoda.tpbot.R;
 import com.novoda.tpbot.bot.service.ServiceConnectionSharedPreferencesPersistence;
 import com.novoda.tpbot.bot.video.calling.VideoCallSharedPreferencesPersistence;
-import com.novoda.tpbot.feature_selection.FeatureSelectionController;
-import com.novoda.tpbot.feature_selection.FeatureSelectionPersistence;
+import com.novoda.tpbot.FeatureSelectionController;
+import com.novoda.tpbot.FeatureSelectionPersistence;
 
 public final class LandingMenuFeatureSelectionController implements FeatureSelectionController<Menu, MenuItem> {
